Rename application to Splintter
