+++ README.md
@@ -1,2 +1,2 @@
-# splinter
+# Splintter
 Android app to split expenses you do between your friends
