+++ app/build.gradle
@@ -12,7 +12,7 @@ android {
     buildToolsVersion "23.0.3"
 
     defaultConfig {
-        applicationId "com.ksss.splinter"
+        applicationId "com.ksss.splintter"
         minSdkVersion 21
         targetSdkVersion 24
         versionCode 1
