+++ app/src/debug/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.ksss.splinter">
+    package="com.ksss.splintter">
 
     <application
         android:name=".DebugApplication"
