+++ app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.ksss.splinter">
+    package="com.ksss.splintter">
 
     <application
         android:name=".MainApplication"
