+++ app/src/main/java/com/ksss/splintter/MainApplication.java
@@ -1,4 +1,4 @@
-package com.ksss.splinter;
+package com.ksss.splintter;
 
 import android.app.Application;
