+++ app/src/main/java/com/ksss/splintter/fend/home/HomeActivity.java
@@ -1,9 +1,9 @@
-package com.ksss.splinter.fend.home;
+package com.ksss.splintter.fend.home;
 
 import android.app.Activity;
 import android.os.Bundle;
 
-import com.ksss.splinter.R;
+import com.ksss.splintter.R;
 
 /**
  * Created by Nahuel Barrios on 7/16/16.
