+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <resources>
-    <string name="app_name">Splinter</string>
+    <string name="app_name">Splintter</string>
     <string name="hello_world">Hello World</string>
 </resources>
